+++ internal/acctest/crypto_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestTLSRSAPrivateKeyPEM(t *testing.T) {
-	key := acctest.TLSRSAPrivateKeyPEM(2048)
+	key := acctest.TLSRSAPrivateKeyPEM(t, 2048)
 
 	if !strings.Contains(key, acctest.PEMBlockTypeRSAPrivateKey) {
 		t.Errorf("key does not contain RSA PRIVATE KEY: %s", key)
@@ -16,8 +16,8 @@ func TestTLSRSAPrivateKeyPEM(t *testing.T) {
 }
 
 func TestTLSRSAPublicKeyPEM(t *testing.T) {
-	privateKey := acctest.TLSRSAPrivateKeyPEM(2048)
-	publicKey := acctest.TLSRSAPublicKeyPEM(privateKey)
+	privateKey := acctest.TLSRSAPrivateKeyPEM(t, 2048)
+	publicKey := acctest.TLSRSAPublicKeyPEM(t, privateKey)
 
 	if !strings.Contains(publicKey, acctest.PEMBlockTypePublicKey) {
 		t.Errorf("key does not contain PUBLIC KEY: %s", publicKey)
@@ -25,10 +25,10 @@ func TestTLSRSAPublicKeyPEM(t *testing.T) {
 }
 
 func TestTLSRSAX509LocallySignedCertificatePEM(t *testing.T) {
-	caKey := acctest.TLSRSAPrivateKeyPEM(2048)
-	caCertificate := acctest.TLSRSAX509SelfSignedCACertificatePEM(caKey)
-	key := acctest.TLSRSAPrivateKeyPEM(2048)
-	certificate := acctest.TLSRSAX509LocallySignedCertificatePEM(caKey, caCertificate, key, "example.com")
+	caKey := acctest.TLSRSAPrivateKeyPEM(t, 2048)
+	caCertificate := acctest.TLSRSAX509SelfSignedCACertificatePEM(t, caKey)
+	key := acctest.TLSRSAPrivateKeyPEM(t, 2048)
+	certificate := acctest.TLSRSAX509LocallySignedCertificatePEM(t, caKey, caCertificate, key, "example.com")
 
 	if !strings.Contains(certificate, acctest.PEMBlockTypeCertificate) {
 		t.Errorf("certificate does not contain CERTIFICATE: %s", certificate)
@@ -36,8 +36,8 @@ func TestTLSRSAX509LocallySignedCertificatePEM(t *testing.T) {
 }
 
 func TestTLSRSAX509SelfSignedCACertificatePEM(t *testing.T) {
-	caKey := acctest.TLSRSAPrivateKeyPEM(2048)
-	caCertificate := acctest.TLSRSAX509SelfSignedCACertificatePEM(caKey)
+	caKey := acctest.TLSRSAPrivateKeyPEM(t, 2048)
+	caCertificate := acctest.TLSRSAX509SelfSignedCACertificatePEM(t, caKey)
 
 	if !strings.Contains(caCertificate, acctest.PEMBlockTypeCertificate) {
 		t.Errorf("CA certificate does not contain CERTIFICATE: %s", caCertificate)
@@ -45,8 +45,8 @@ func TestTLSRSAX509SelfSignedCACertificatePEM(t *testing.T) {
 }
 
 func TestTLSRSAX509SelfSignedCertificatePEM(t *testing.T) {
-	key := acctest.TLSRSAPrivateKeyPEM(2048)
-	certificate := acctest.TLSRSAX509SelfSignedCertificatePEM(key, "example.com")
+	key := acctest.TLSRSAPrivateKeyPEM(t, 2048)
+	certificate := acctest.TLSRSAX509SelfSignedCertificatePEM(t, key, "example.com")
 
 	if !strings.Contains(certificate, acctest.PEMBlockTypeCertificate) {
 		t.Errorf("certificate does not contain CERTIFICATE: %s", certificate)
@@ -54,7 +54,7 @@ func TestTLSRSAX509SelfSignedCertificatePEM(t *testing.T) {
 }
 
 func TestTLSRSAX509CertificateRequestPEM(t *testing.T) {
-	csr, key := acctest.TLSRSAX509CertificateRequestPEM(2048, "example.com")
+	csr, key := acctest.TLSRSAX509CertificateRequestPEM(t, 2048, "example.com")
 
 	if !strings.Contains(csr, acctest.PEMBlockTypeCertificateRequest) {
 		t.Errorf("certificate does not contain CERTIFICATE REQUEST: %s", csr)